Add window event listener 'beforeinstallprompt' to prevent default of immediate banner install; we want to install the banner at a later point of time
On feed.js we check to see if browser has fired the event listener because we cannot do this on our own.  If it has, then we preventDefault() and catch the event in app.js variable promptDeferment and fire it (promptDeferment.prompt(), in feed.js when button with plus icon is clicked.