In the Service Worker, comment out the Pre-Caching fetch event and replace it with a Dynamic Caching fetch event
Like the Pre-Caching fetch event, assets are fetched first from the static cache and if they are not found in the static cache, then the asset is fetched from the Network.  Except in this case, if the asset is fetched from the network, it is put into a dynamic cache for a better offline-first experience.