Implement a Cache, then Network with Dynamic Caching Strategy for GET request for url which triggers json file and kicks off the createCard function, and use the Dynamic Caching with Network Fallback and Offline Fallback Page Strategy for all other assets
This ensures offline-first capabilities for dynamic card creation in feed.js and that static caching along with dynamic caching is used for all other assets for offline first capabilities.  Also, if it is the case where a cache response (both static and dynamic) and a network response fails, an offline html page will give the option to direct the user back to the main index page which has been pre-cached in the install event.