Create a function to initialize media which is called when the user opens modal to create a post
I'm going to build this one step at a time.  In this commit, I check to
see if the Browser supports the mediaDevices API interface which handles
video and audio.  An image is considered video.  If the Browser does not
support mediaDevices then I just create the property mediaDevices and set
it to an empty object.  The goal here is to bring some of the older
Browsers into the newer syntax, not to recreate mediaDevices.